finish

+++ app/views/inventory/index.php
@@ -6,7 +6,7 @@
         </button>
     </nav>
     <nav class="navbar navbar-light bg-light">
-        <form class="form-inline" method="POST" action="<?=BASEURL?>public/inventory/search">
+        <form class="form-inline" method="POST" action="<?= BASEURL ?>public/inventory/search">
             <input name="searchProduct" class="form-control mr-sm-2" type="search" placeholder="Search By Id or Name" aria-label="Search">
             <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
         </form>
@@ -26,6 +26,12 @@
         <tbody>
             <?php
             $id = 0;
+            $batas = 4;
+            $halaman = $data['halaman'];
+            $halaman_awal = $data['halaman_awal'];
+            $total_halaman = ceil($data['dataRange'] / $batas);
+            $previous = $halaman - 1;
+            $next = $halaman + 1;
             foreach ($data['inventory'] as $product) :
             ?>
                 <tr>
@@ -35,14 +41,14 @@
                     <td><?= $product['P_sell'] ?></td>
                     <td><?= $product['P_price'] ?></td>
                     <td class="justify-content-between">
-                        <button class=" btn btn-md btn-success" data-toggle="modal" data-target="#formEditModal<?=$id?>" >Edit</button>
-                        <button class=" btn btn-md btn-danger" data-toggle="modal" data-target="#formDetailModal<?=$id?>" >
+                        <button class=" btn btn-md btn-success" data-toggle="modal" data-target="#formEditModal<?= $id ?>">Edit</button>
+                        <button class=" btn btn-md btn-danger" data-toggle="modal" data-target="#formDetailModal<?= $id ?>">
                             Detail</button>
                     </td>
                 </tr>
 
                 <!-- Modal For Edit-->
-                <div class="modal fade" id="formEditModal<?=$id?>" tabindex="-1" role="dialog" aria-labelledby="formEditModalLabel" aria-hidden="true">
+                <div class="modal fade" id="formEditModal<?= $id ?>" tabindex="-1" role="dialog" aria-labelledby="formEditModalLabel" aria-hidden="true">
                     <div class="modal-dialog" role="document">
                         <div class="modal-content">
                             <div class="modal-header">
@@ -52,7 +58,7 @@
                                 </button>
                             </div>
 
-                            <form action="<?= BASEURL; ?>public/inventory/edit/<?=$product['P_Id']?>" method="post">
+                            <form action="<?= BASEURL; ?>public/inventory/edit/<?= $product['P_Id'] ?>" method="post">
                                 <div class="modal-body">
                                     <div class="form-group">
                                         <label for="pname">Product Name</label>
@@ -84,8 +90,7 @@
                                     </div>
                                 </div>
                                 <div class="modal-footer">
-                                    <a type="button" href="<?= BASEURL; ?>public/inventory/delete/<?= $product['P_Id'] ?>" class="btn btn-danger" 
-                                    onclick="return confirm('Yakin Menghapus Data Produk Ini?')">Delete</a>
+                                    <a type="button" href="<?= BASEURL; ?>public/inventory/delete/<?= $product['P_Id'] ?>" class="btn btn-danger" onclick="return confirm('Yakin Menghapus Data Produk Ini?')">Delete</a>
                                     <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
                                     <button type="submit" class="btn btn-primary">Save changes</button>
                                 </div>
@@ -96,7 +101,7 @@
 
                 <!--Modal for detail -->
 
-                <div class="modal fade" id="formDetailModal<?=$id?>" tabindex="-1" role="dialog" aria-labelledby="formDetailModalLabel" aria-hidden="true">
+                <div class="modal fade" id="formDetailModal<?= $id ?>" tabindex="-1" role="dialog" aria-labelledby="formDetailModalLabel" aria-hidden="true">
                     <div class="modal-dialog" role="document">
                         <div class="modal-content">
                             <div class="modal-header">
@@ -106,41 +111,41 @@
                                 </button>
                             </div>
 
-                    
-                                <div class="modal-body">
-                                    <div class="form-group">
-                                        <label for="pname">Product Name</label>
-                                        <input readonly type="text" value="<?= $product['P_name'] ?>" class="form-control" id="pname" name="P_name" placeholder="nama produk">
-                                    </div>
-                                    <div class="form-group">
-                                        <label for="mfdate">Manufacture Date</label>
-                                        <input readonly  type="date" value="<?= $product['P_mfdate'] ?>" class="form-control" id="mfdate" name="P_mfdate" placeholder="Manufacture date">
-                                    </div>
-                                    <div class="form-group">
-                                        <label for="expdate">Expire Date</label>
-                                        <input readonly type="date" value="<?= $product['P_expdate'] ?>" class="form-control" id="expdate" name="P_expdate" placeholder="Exp date">
-                                    </div>
-                                    <div class="form-group">
-                                        <label for="price">Price</label>
-                                        <input readonly type="number" value="<?= $product['P_price'] ?>" class="form-control" id="price" name="price" placeholder="price">
-                                    </div>
-                                    <div class="form-group">
-                                        <label for="splier">Supplier</label>
-                                        <input readonly type="text" value="<?= $product['P_suplier'] ?>" class="form-control" id="splier" name="P_suplier" placeholder="Suplier name">
-                                    </div>
-                                    <div class="form-group">
-                                        <label for="sellp">Sell price</label>
-                                        <input readonly type="number" value="<?= $product['P_sell'] ?>" class="form-control" id="sellp" name="P_sell" placeholder="sell price">
-                                    </div>
-                                    <div class="form-group">
-                                        <label for="sellp">Stock</label>
-                                        <input readonly type="number" value="<?= $product['P_stock'] ?>" class="form-control" id="pstock" name="P_stock" placeholder="Stock">
-                                    </div>
+
+                            <div class="modal-body">
+                                <div class="form-group">
+                                    <label for="pname">Product Name</label>
+                                    <input readonly type="text" value="<?= $product['P_name'] ?>" class="form-control" id="pname" name="P_name" placeholder="nama produk">
                                 </div>
-                                <div class="modal-footer">
-                                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                                <div class="form-group">
+                                    <label for="mfdate">Manufacture Date</label>
+                                    <input readonly type="date" value="<?= $product['P_mfdate'] ?>" class="form-control" id="mfdate" name="P_mfdate" placeholder="Manufacture date">
                                 </div>
-                
+                                <div class="form-group">
+                                    <label for="expdate">Expire Date</label>
+                                    <input readonly type="date" value="<?= $product['P_expdate'] ?>" class="form-control" id="expdate" name="P_expdate" placeholder="Exp date">
+                                </div>
+                                <div class="form-group">
+                                    <label for="price">Price</label>
+                                    <input readonly type="number" value="<?= $product['P_price'] ?>" class="form-control" id="price" name="price" placeholder="price">
+                                </div>
+                                <div class="form-group">
+                                    <label for="splier">Supplier</label>
+                                    <input readonly type="text" value="<?= $product['P_suplier'] ?>" class="form-control" id="splier" name="P_suplier" placeholder="Suplier name">
+                                </div>
+                                <div class="form-group">
+                                    <label for="sellp">Sell price</label>
+                                    <input readonly type="number" value="<?= $product['P_sell'] ?>" class="form-control" id="sellp" name="P_sell" placeholder="sell price">
+                                </div>
+                                <div class="form-group">
+                                    <label for="sellp">Stock</label>
+                                    <input readonly type="number" value="<?= $product['P_stock'] ?>" class="form-control" id="pstock" name="P_stock" placeholder="Stock">
+                                </div>
+                            </div>
+                            <div class="modal-footer">
+                                <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                            </div>
+
                         </div>
                     </div>
                 </div>
@@ -150,6 +155,24 @@
             ?>
         </tbody>
     </table>
+
+  <nav>
+			<ul class="pagination justify-content-center">
+				<li class="page-item">
+					<a class="page-link" <?php if($halaman > 1){ echo "href=".BASEURL."public/inventory/index/$previous"; } ?>>Previous</a>
+				</li>
+				<?php 
+				for($x=1;$x<=$total_halaman;$x++){
+					?> 
+					<li class="page-item"><a class="page-link" href="<?php echo BASEURL."public/inventory/index/$x" ?>"><?php echo $x; ?></a></li>
+					<?php
+				}
+				?>				
+				<li class="page-item">
+					<a  class="page-link" <?php if($halaman < $total_halaman) { echo "href=".BASEURL."public/inventory/index/$next"; } ?>>Next</a>
+				</li>
+			</ul>
+		</nav>
 
     <!-- Modal -->
     <div class="modal fade" id="formModal" tabindex="-1" role="dialog" aria-labelledby="formModalLabel" aria-hidden="true">
